% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.SDMfit.R
\name{coef.SDMfit}
\alias{coef.SDMfit}
\title{Gets regression coefficients from a local model, i.e. a SDMfit object.}
\usage{
\method{coef}{SDMfit}(object, standardise = F, level = 0.95, ...)
}
\arguments{
\item{object}{A SDMfit object, typically obtained with trophicSDM() and available in the field $model of a trophicSDMfit object}

\item{standardise}{Whether to standardise regression coefficients. Default to FALSE. If TRUE, coefficients are standardised using the latent variable standardisation (see Grace et al. 2018) for more details.}

\item{level}{The confidence level of credible intervals, only available for stan_glm method. Default to 0.95.}

\item{...}{additional arguments}
}
\value{
A table containing the inferred coefficients (with credible intervals or p-values when available).
}
\description{
Gets regression coefficients (eventually standardised) of a local model, i.e. a SDMfit object. p-values or credible intervales are returned when available.
}
\examples{

}
\references{
Grace, J. B., Johnson, D. J., Lefcheck, J. S., and Byrnes, J. E. K.. 2018. Quantifying relative importance: computing standardized effects in models with binary outcomes. Ecosphere 9(6):e02283.
}
\author{
Giovanni Poggiato
}
