% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeVariableImportance.R
\name{computeVariableImportance}
\alias{computeVariableImportance}
\title{Computes variable importance of (groups of variable) for each species of a trophicSDM model.
Variable importance are computed as the standardised regression coefficients (summed across species of the same group). Standardisation is done using latent variable standardisation described in Grace et al. 2018.}
\usage{
computeVariableImportance(tSDM, groups = NULL)
}
\arguments{
\item{tSDM}{A trophicSDMfit object obtained with trophicSDM()}

\item{groups}{A list where each element is group. Each group is specified as a vector containing species or environmental covariates names of a given group. Each element of the list (i.e. each group) has to be named.}
}
\value{
A groups x species matrix containing variable importance for each groups of variables and each species.
}
\description{
Computes variable importance of (groups of variable) for each species of a trophicSDM model.
Variable importance are computed as the standardised regression coefficients (summed across species of the same group). Standardisation is done using latent variable standardisation described in Grace et al. 2018.
}
\examples{

}
\author{
Giovanni Poggiato
#' @references Grace, J. B., Johnson, D. J., Lefcheck, J. S., and Byrnes, J. E. K.. 2018. Quantifying relative importance: computing standardized effects in models with binary outcomes. Ecosphere 9(6):e02283.
}
