% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateModelFit.R
\name{evaluateModelFit}
\alias{evaluateModelFit}
\title{Evaluates goodness of fit by comparing a true versus a predicted dataset of species distribution
Ypredicted is tipically predicted using a prediction method of trophicSDM (in cross validation or not)}
\usage{
evaluateModelFit(tSDM, Ynew = NULL, Ypredicted = NULL)
}
\arguments{
\item{tSDM}{A trophicSDMfit object obtained with trophicSDM()}

\item{Ynew}{A sites x species matrix containing the true species occurrences state}

\item{Ypredicted}{A sites x species matrix containing the predicted species occurrences state.}
}
\value{
A table specifying the goodness of fit metrics for each species. For presence-absence data, the model computes TSS and AUC. For Gaussian data, the R2.
}
\description{
Evaluates goodness of fit by comparing a true versus a predicted dataset of species distribution
Ypredicted is tipically predicted using a prediction method of trophicSDM (in cross validation or not)
}
\examples{

}
\references{
Grace, J. B., Johnson, D. J., Lefcheck, J. S., and Byrnes, J. E. K.. 2018. Quantifying relative importance: computing standardized effects in models with binary outcomes. Ecosphere 9(6):e02283.
}
\author{
Giovanni Poggiato
}
