% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.trophicSDMfit.R
\name{loo.trophicSDMfit}
\alias{loo.trophicSDMfit}
\title{Computes an approximation of loo for the whole model
The global loo is computed by summing the loo of all the local models (since the likelihood factorises, the log-likelihood can be summed)}
\usage{
\method{loo}{trophicSDMfit}(tSDM)
}
\arguments{
\item{tSDM}{A trophicSDMfit object obtained with trophicSDM()}
}
\description{
Computes an approximation of loo for the whole model
The global loo is computed by summing the loo of all the local models (since the likelihood factorises, the log-likelihood can be summed)
}
\examples{

}
\author{
Giovanni Poggiato
}
