% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SDMfit.R
\name{plot.SDMfit}
\alias{plot.SDMfit}
\title{Plots the regression coeffiecients of a local SDMfit model}
\usage{
\method{plot}{SDMfit}(x, level = 0.95, ...)
}
\arguments{
\item{x}{A SDMfit object, typically obtained with trophicSDM() and available in the field $model of a trophicSDMfit object}

\item{level}{the confidence level of the confidence intervals}

\item{...}{additional arguments}
}
\description{
Plots the regression coeffiecients of a local SDMfit model
}
\examples{

}
\author{
Giovanni Poggiato
}
