% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.trophicSDMfit.R
\name{plot.trophicSDMfit}
\alias{plot.trophicSDMfit}
\title{Plots the regression coefficients of a fitted trophicSDM model}
\usage{
\method{plot}{trophicSDMfit}(x, species = NULL, ...)
}
\arguments{
\item{x}{A trophicSDMfit object obtained with trophicSDM()}

\item{species}{A vector of species names to be plot. If NULL (default), all species are plotted.}

\item{...}{additional arguments}
}
\description{
Plots the regression coefficients of a fitted trophicSDM model
}
\examples{

}
\author{
Giovanni Poggiato
}
