% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotG.R
\name{plotG}
\alias{plotG}
\title{Plot the metaweb G
Plots the metaweb G used to fit the trophicSDM model}
\usage{
plotG(tSDM)
}
\arguments{
\item{tSDM}{A trophicSDMfit object obtained with trophicSDM()}
}
\value{
A ggnet object
}
\description{
Plot the metaweb G
Plots the metaweb G used to fit the trophicSDM model
}
\examples{

}
\author{
Giovanni Poggiato
}
