% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotG_inferred.R
\name{plotG_inferred}
\alias{plotG_inferred}
\title{Plot the metaweb G with links colored accordingly to the inferred prey-predator regression coefficients of a fitted trophicSDM model.
Plots the metaweb G, where each predator-prey link is colored according to whether the related regression coefficient if inferred as positive (in red), negative (in blue) or non-significant (dashed grey line) according to the confidence level specified in "level". Estimates of the significant standardised regression coefficients are pasted on the links. Only works if species are modelled as a function of their preys or predators without composite variables (i.e., if tSDM is fitted with sp.formula = NULL and sp.partition = NULL)}
\usage{
plotG_inferred(tSDM, level = 0.9)
}
\arguments{
\item{tSDM}{A trophicSDMfit object obtained with trophicSDM()}

\item{level}{The confidence level used to decide whether regression coefficients are non-significant or not. Default to 0.9.}
}
\value{
A ggnet object.
}
\description{
Plot the metaweb G with links colored accordingly to the inferred prey-predator regression coefficients of a fitted trophicSDM model.
Plots the metaweb G, where each predator-prey link is colored according to whether the related regression coefficient if inferred as positive (in red), negative (in blue) or non-significant (dashed grey line) according to the confidence level specified in "level". Estimates of the significant standardised regression coefficients are pasted on the links. Only works if species are modelled as a function of their preys or predators without composite variables (i.e., if tSDM is fitted with sp.formula = NULL and sp.partition = NULL)
}
\examples{

}
\author{
Giovanni Poggiato
}
