% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.SDMfit.R
\name{predict.SDMfit}
\alias{predict.SDMfit}
\title{Computes predicted values for a local model, i.e., a fitted SDMfit object
This is sequentially called, for each species, by the function trophicSDM.predict}
\usage{
\method{predict}{SDMfit}(object, newdata, pred_samples, prob.cov, ...)
}
\arguments{
\item{object}{A SDMfit object, typically obtained with trophicSDM() and available in the field $model of a trophicSDMfit object}

\item{newdata}{A matrix containing both environmental covariates and the biotic variables that the local model uses to predict the species distribution.}

\item{pred_samples}{Number of samples to draw from species posterior predictive distribution when method = "stan_glm". If NULL, set by the default to the number of iterations/10.}

\item{prob.cov}{If set to FALSE, it gives back also predicted presence-absences (which is then used by trophicSDM.predict to predict the predators).}

\item{...}{additional arguments}
}
\value{
A list containing for each species the predicted value at each sites. If method = "stan_glm", then each element of the list is a sites x pred_samples matrix containing the posterior predictive distribution of the species at each sites.
}
\description{
Computes predicted values for a local model, i.e., a fitted SDMfit object
This is sequentially called, for each species, by the function trophicSDM.predict
}
\examples{

}
\author{
Giovanni Poggiato and Jérémy Andréoletti
}
