% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.trophicSDMfit.R
\name{predict.trophicSDMfit}
\alias{predict.trophicSDMfit}
\title{Computes predicted values from the fitted trophicSDMfit model}
\usage{
\method{predict}{trophicSDMfit}(
  object,
  Xnew = NULL,
  prob.cov = F,
  pred_samples = NULL,
  run.parallel = T,
  verbose = F,
  fullPost = T,
  filter.table = NULL,
  ...
)
}
\arguments{
\item{object}{A trophicSDMfit object obtained with trophicSDM()}

\item{Xnew}{a matrix specifying the environmental covariates for the predictions to be made. If NULL (default), predictions are done on the training dataset (e.g. by setting Xnew = tSDM$data$X).}

\item{prob.cov}{Parameter to predict with trophicSDM with presence-absence data. Whether to use predicted probability of presence (prob.cov = T) or the transformed presence-absences (default, prov.cov = F) to predict species distribution.}

\item{pred_samples}{Number of samples to draw from species posterior predictive distribution when method = "stan_glm". If NULL, set by the default to the number of iterations/10.}

\item{run.parallel}{Whether to use parallelise code when possible. Can speed up computation time.}

\item{verbose}{Whether to print advances of the algorithm}

\item{fullPost}{Optional parameter for stan_glm only. Whether to give back the full posterior predictive distribution (default, fillPost = TRUE) or just the posterior mean, and 2.5\% and 97.5\% quantiles,}

\item{filter.table}{Optional, default to NULL, should be provided only if the users wants to filter some species predictions. A sites x species matrix of zeros and ones.}

\item{...}{additional arguments}
}
\value{
A list containing for each species the predicted value at each sites. If method = "stan_glm", then each element of the list is a sites x pred_samples matrix containing the posterior predictive distribution of the species at each sites.
}
\description{
Computes predicted values from the fitted trophicSDMfit model at environmental conditions specified by \code{Xnew}. Once predictions have been obtained, their quality can eventually be evaluated with \code{evaluateModelFit()}.
}
\examples{

}
\author{
Giovanni Poggiato and Jérémy Andréoletti
}
