% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFundamental.R
\name{predictFundamental}
\alias{predictFundamental}
\title{Predicts species fundamental niche}
\usage{
predictFundamental(
  tSDM,
  Xnew = NULL,
  pred_samples = NULL,
  run.parallel = T,
  verbose = F,
  fullPost = T
)
}
\arguments{
\item{tSDM}{A trophicSDMfit object obtained with trophicSDM()}

\item{Xnew}{a matrix specifying the environmental covariates for the predictions to be made. If NULL (default), predictions are done on the training dataset (e.g. by setting Xnew = tSDM$data$X).}

\item{pred_samples}{Number of samples to draw from species posterior predictive distribution when method = "stan_glm". If NULL, set by the default to the number of iterations/10.}

\item{run.parallel}{Whether to use parallelise code when possible. Can speed up computation time.}

\item{verbose}{Whether to print advances of the algorithm.}

\item{fullPost}{Optional parameter for stan_glm only. Whether to give back the full posterior predictive distribution (default, fillPost = TRUE) or just the posterior mean, and 2.5\% and 97.5\% quantiles.}
}
\value{
A list containing for each species the predicted value at each sites. If method = "stan_glm", then each element of the list is a sites x pred_samples matrix containing the posterior predictive distribution of the species at each sites.
}
\description{
Compues predicted values of the fundamental niches of species from the fitted trophicSDMfit model at environmental conditions specified by \code{Xnew}. Predictions are obtained by setting preys to present when mode = "prey" or setting predators to absent when mode = "predator".
}
\examples{

}
\author{
Giovanni Poggiato and Jérémy Andréoletti
}
