% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SDMfit.R
\name{print.SDMfit}
\alias{print.SDMfit}
\title{Prints a SDMfit object}
\usage{
\method{print}{SDMfit}(x, ...)
}
\arguments{
\item{x}{A SDMfit object, typically obtained with trophicSDM() and available in the field $model of a trophicSDMfit object}

\item{...}{additional arguments}
}
\description{
Prints a SDMfit object
}
\examples{

}
\author{
Giovanni Poggiato
}
