% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trophicSDMfit.R
\name{print.trophicSDMfit}
\alias{print.trophicSDMfit}
\title{Prints a fitted trophicSDM model}
\usage{
\method{print}{trophicSDMfit}(x, ...)
}
\arguments{
\item{x}{A trophicSDMfit object obtained with trophicSDM()}

\item{...}{additional arguments}
}
\description{
Prints a fitted trophicSDM model
}
\examples{

}
\author{
Giovanni Poggiato
}
