% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SDMfit.R
\name{summary.SDMfit}
\alias{summary.SDMfit}
\title{Summary of a fitted SDMfit model}
\usage{
\method{summary}{SDMfit}(object, ...)
}
\arguments{
\item{object}{A SDMfit object, typically obtained with trophicSDM() and available in the field $model of a trophicSDMfit object}

\item{...}{additional arguments}
}
\description{
Summary of a fitted SDMfit model
}
\examples{

}
\author{
Giovanni Poggiato
}
