% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.trophicSDMfit.R
\name{summary.trophicSDMfit}
\alias{summary.trophicSDMfit}
\title{Prints the summary of a fitted trophicSDM model}
\usage{
\method{summary}{trophicSDMfit}(object, ...)
}
\arguments{
\item{object}{A trophicSDMfit object obtained with trophicSDM()}

\item{...}{additional arguments}
}
\description{
Prints the summary of a fitted trophicSDM model
}
\examples{

}
\author{
Giovanni Poggiato
}
