% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trocSDM-package.R
\docType{package}
\name{trocSDM}
\alias{trocSDM}
\title{trocSDM.}
\description{
Package to fit a trophic Species Distribution Model, analyse it and predict. See Poggiato et al. In prep.
}
\author{
Giovanni Poggiato \email{giov.poggiato@gmail.com} and Jérémy Andréoletti
}
