% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophicSDM.R
\name{trophicSDM}
\alias{trophicSDM}
\title{Fitting a trophic Species distribution model}
\usage{
trophicSDM(
  Y,
  X,
  G,
  env.formula = NULL,
  sp.formula = NULL,
  sp.partition = NULL,
  penal = NULL,
  mode = "prey",
  method = "stan_glm",
  family,
  iter = 1000,
  run.parallel = TRUE,
  chains = 2,
  verbose = F
)
}
\arguments{
\item{Y}{The sites x species matrix containing observed species distribution (e.g. presence-absence).}

\item{X}{The design matrix, i.e. sites x predictor matrix containing the value of each explanatory variable (e.g. the environmental conditions) at each site.}

\item{G}{The species interaction network (aka metaweb). Needs to be an igraph object. Links must go from predator to preys. It needs to be a directed acyclic graph.}

\item{env.formula}{A list that contains for each species the formula that describes the abiotic part of the model. The names of the list must coincide with the names of the species. The details of model specification are given under ‘Details’.}

\item{sp.formula}{(optional) It allows to specify a particular definition of the biotic part of the model, e.g., using composite variables (e.g., richness), or an interaction of the biotic and abitic component. More details in 'Details'.}

\item{sp.partition}{(optional) a list to specify groups of species that are used to compute composite variables, e.g., a species can be modelled as a function of the richness of each group of preys. It has to be a list, each element is a vector containing the names of species in the group.}

\item{penal}{(optional, default to NULL) Penalisation method to shrink regression coefficients.If NULL (default), the model does not penalise the regression coefficient. For now, available penalisation method are "horshoe" for stan_glm, "elasticnet" for glm and  "coeff.signs" (prey coefficients are set to posite and predator coefficients to negative) for glm and stan_glm.}

\item{mode}{"prey" if bottom-up control (default), "predators" otherwise. Notice that G needs to be such that links point from predators to prey.}

\item{method}{which SDM method to use. For now the available choises are: "glm" (frequentist) or "stan_glm" (full bayesian MCMC, default). Notice that using "glm" does not allow error propagation when predicting.}

\item{family}{the family parameter of the glm function (see glm). family=gaussian for gaussian data or family=binomial(link = "logit") or binomial(link = "probit") for presence-absence data.}

\item{iter}{(for method="stan_glm" only) Number of iterations for each MCMC chain if stan_glm is used}

\item{run.parallel}{Whether species models are fitted in parallel (can speed computational up time). Default to FALSE.}

\item{chains}{(for method="stan_glm" only) Number of MCMC chains (default to 2)}

\item{verbose}{Whether to print algorithm progresses}
}
\value{
A "trophicSDMfit" object, containing:
\item{model}{A list containing the local models (i.e. a SDM for each species). Each local model is an object of class "SDMfit". See ?SDMfit for more informations.}
\item{Y}{A numeric vector of standard errors on parameters}

\item{form.all}{A list describing each species formula (both biotic and abiotic terms)}

\item{data}{A list containing all the data used to fit the model}

\item{model.call}{A list containing the modeling choices of the fitted model (e.g. method, penalisation...)}

\item{coef}{A list containing, for each species, the inferred coefficients (with credible intervals or p-values when available)}
\item{MCMC.diag}{MCMC convergence metrics, only available for MCMC methods}

\item{AIC}{Model's AIC}

\item{log.lik}{Model's log.likelihood}
}
\description{
trophicSDM is used to fit a trophic species distribution model. Requires the species distribution data Y (the sites x species matrix), explanatory variables X and a directed acyclic graph G containing species interactions (i.e., the metaweb, with links going from predators to prey). The function fits the distribution of each species as a function of their preys (with mode = "prey", by default) or predators (if set mode = "predator").
}
\details{
"sp.formula" and "sp.partition" can be combined to define any kind of composite variables for the biotic part of the formula. "sp.formula" can be :
\itemize{
\item A string defining a formula as function of "richness". E.g., sp.formula="richness+I(richness)^2" (species are modelled as a function of a quadratic polyome of their prey richness), "I(richness>0)" (species are modelled as a function of a dummy variable that is equal to 1 when at least one species is present). Importantly, when group of preys (or predators) are specified by "sp.partition", species are modeled as a function of the composite variable specified by "sp.formula" for each of their prey groups.
\item A more flexible option is to specify sp.formula as a list (whose names are species' names) that contains for each species the definition biotic part of the model. Notice that, in this case, the function does not check that the model is a DAG. This allow to define any kind of composite variable, or to model interactions between environmental covariates and preys (or predators).
}
}
\examples{
data(Y)
data(X)
data(G)
env.formula = as.list(rep("~ X_1 + X_2", ncol(Y)))
names(env.formula) = colnames(Y)
trophicSDM(Y,X,G, env.formula, family = binomial(link = logit()), mode = "prey", method = "stan_glm")
}
\author{
Giovanni Poggiato and Jérémy Andréletti
}
