% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophicSDM_CV.R
\name{trophicSDM_CV}
\alias{trophicSDM_CV}
\title{Compute K-fold cross-validation predicted values from a fitted trophicSDM model}
\usage{
trophicSDM_CV(
  tSDM,
  K,
  partition = NULL,
  prob.cov = F,
  pred_samples = NULL,
  iter = NULL,
  chains = NULL,
  run.parallel = T,
  verbose = F
)
}
\arguments{
\item{tSDM}{A trophicSDMfit object obtained with trophicSDM()}

\item{K}{The number of folds for the K-fold cross validation}

\item{partition}{Optional parameter. A partition vector to specify a partition in K fold for cross validation}

\item{prob.cov}{Parameter to predict with trophicSDM with presence-absence data. Whether to use predicted probability of presence (prob.cov = T) or the transformed presence-absences (default, prov.cov = F) to predict species distribution.}

\item{pred_samples}{Number of samples to draw from species posterior predictive distribution when method = "stan_glm". If NULL, set by the default to the number of iterations/10.}

\item{iter}{For method = "stan_glm": number of iterations of each MCMC chains to fit the trophicSDM model. Default to the number of iterations used to fit the provided trophicSDMfit object}

\item{chains}{For method = "stan_glm": number of MCMC chains to fit the trophicSDM model. Default to the number of iterations used to fit the provided trophicSDMfit object}

\item{run.parallel}{Whether to use parallelise code when possible. Default to TRUE. Can speed up computation time}

\item{verbose}{Whether to print advances of the algorithm}
}
\value{
A list containing:
\item{meanPred}{a sites x species matrix of predicted occurrences of species for each site (e.g. probability of presence). With stan_glm the posterior predictive mean is return}
\item{Pred975,Pred025}{Only for method = "stan_glm", the 97.5\% and 2.5\% quantiles of the predictive posterior distribution}
\item{partition}{the partition vector used to compute the K fold cross-validation}
}
\description{
Once the CV predicted values are obtained, their quality can be evaluated with \code{evaluateModelFit()}.
}
\examples{

}
\author{
Giovanni Poggiato
}
