% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulData.R
\docType{data}
\name{X}
\alias{X}
\title{Simulated environemntal covariates X}
\format{
A site x covariates matrix X
}
\usage{
data(X)
}
\description{
Simulated environemntal covariates X
}
\examples{
data(X)
}
\author{
Giovanni Poggiato
}
\keyword{#'}
\keyword{datasets}
