% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulData.R
\docType{data}
\name{Y}
\alias{Y}
\title{Simulated species distribution Y}
\format{
A site x species matrix Y, a site x covariates matrix X and a trophic interaction network G (object igraph)
}
\usage{
data(Y)
}
\description{
Simulated species distribution Y
}
\examples{
data(Y)
}
\author{
Giovanni Poggiato
}
\keyword{datasets}
